espresso 25.9
The `espresso` zip URL has changed with this new release, so this
updates the cask accordingly. Besides that, this updates the
`livecheck` block URL to align with the download URL on the homepage
(it's basically the same content as the `/updates/` page) and to make
the regex a bit more flexible about matching the attribute value.